Group objects now contain group_code. Added PATCH to generate a new group code. avg_user_rating is now a float

+++ MovieNightAPI/Models/GroupMovieRating.cs
@@ -11,6 +11,6 @@ namespace MovieNightAPI.Models
         public int group_id { get; set; }
         public int tmdb_movie_id { get; set; }
         public int user_rating { get; set; }
-        public double avg_user_rating { get; set; }
+        public float avg_user_rating { get; set; }
     }
 }
